Refactor fetchTeam
