+++ client/src/App.jsx
@@ -1,5 +1,4 @@
-import React, { useContext, useEffect, Suspense } from 'react';
-import axios from 'axios';
+import React, { Suspense } from 'react';
 import './App.scss';
 import { Router, View } from 'react-navi';
 import HelmetProvider from 'react-navi-helmet';
@@ -11,8 +10,7 @@ import { initAll } from 'govuk-frontend';
 import Layout from './components/layout';
 import routes from './routes';
 import ApplicationSpinner from './components/ApplicationSpinner';
-import { useAxios, useIsMounted } from './utils/hooks';
-import { TeamContext } from './utils/TeamContext';
+import { useFetchTeam } from './utils/hooks';
 
 if (window.ENVIRONMENT_CONFIG) {
   // eslint-disable-next-line no-console
@@ -46,48 +44,8 @@ const RouterView = () => {
   const { t } = useTranslation();
   const [keycloak, initialized] = useKeycloak();
 
-  const axiosInstance = useAxios();
-  const isMounted = useIsMounted();
-  const { setTeam } = useContext(TeamContext);
-
   initAll();
-
-  useEffect(() => {
-    const source = axios.CancelToken.source();
-    if (initialized) {
-      const {
-        tokenParsed: { team_id: teamid },
-      } = keycloak;
-      const fetchData = async () => {
-        if (axiosInstance) {
-          try {
-            if (teamid) {
-              const response = await axiosInstance.get(
-                `/refdata/v2/entities/team?filter=id=eq.${teamid}`,
-                {
-                  cancelToken: source.token,
-                }
-              );
-              if (isMounted.current) {
-                const [team] = response.data.data;
-                setTeam(team);
-              }
-            } else {
-              // TODO: Redirect the user here because they have no teamid in KC...
-            }
-          } catch (error) {
-            if (isMounted.current) {
-              setTeam({});
-            }
-          }
-        }
-      };
-      fetchData();
-    }
-    return () => {
-      source.cancel('Cancelling request');
-    };
-  }, [axiosInstance, initialized, isMounted, keycloak, setTeam]);
+  useFetchTeam();
 
   return initialized ? (
     <Router
