+++ client/src/utils/hooks.js
@@ -5,6 +5,7 @@ import { useKeycloak } from '@react-keycloak/web';
 import { useCurrentRoute } from 'react-navi';
 import Logger from './logger';
 import { AlertContext } from './AlertContext';
+import { TeamContext } from './TeamContext';
 
 export const useAxios = () => {
   const [keycloak, initialized] = useKeycloak();
@@ -64,6 +65,47 @@ export const useIsMounted = () => {
     };
   }, []);
   return isMounted;
+};
+
+export const useFetchTeam = () => {
+  const [keycloak, initialized] = useKeycloak();
+  const axiosInstance = useAxios();
+  const isMounted = useIsMounted();
+  const { setTeam } = useContext(TeamContext);
+  useEffect(() => {
+    const source = axios.CancelToken.source();
+    if (initialized) {
+      const {
+        tokenParsed: { team_id: teamid },
+      } = keycloak;
+      const fetchData = async () => {
+        try {
+          if (teamid) {
+            const response = await axiosInstance.get(
+              `refdata/v2/entities/team?filter=id=eq.${teamid}`,
+              {
+                cancelToken: source.token,
+              }
+            );
+            if (isMounted.current) {
+              const team = response.data.data[0];
+              setTeam(team);
+            }
+          } else {
+            // TODO: Redirect the user here because they have no teamid in KC...
+          }
+        } catch (error) {
+          if (isMounted.current) {
+            setTeam({});
+          }
+        }
+      };
+      fetchData();
+    }
+    return () => {
+      source.cancel('Cancelling request');
+    };
+  }, [axiosInstance, initialized, isMounted, keycloak, setTeam]);
 };
 
 export default useIsMounted;
